Updated to HDRP. Skybox, fog, lighting changes
Turns out that with nice graphics the infinite staircase illsuion sort of breaks down meaning we should probably just stick to non illusiony shit. Either that or we have it look very different when entering fucko illusion puzzle area but that would have to be done outside of game jam time.